wasm/webapi/esm-integration/source-phase.tentative.html

<!doctype html>
<title>Source phase imports</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });

import source exportedNamesSource from "./resources/exported-names.wasm";

assert_equals(typeof AbstractModuleSource, "undefined");

{
  const AbstractModuleSource = Object.getPrototypeOf(WebAssembly.Module).constructor;
  assert_equals(AbstractModuleSource.name, "AbstractModuleSource");
  assert_not_equals(AbstractModuleSource, Function);

  const AbstractModuleSourceProto = Object.getPrototypeOf(WebAssembly.Module.prototype).constructor;
  assert_not_equals(AbstractModuleSourceProto, Object);

  assert_true(exportedNamesSource instanceof WebAssembly.Module);
  assert_true(exportedNamesSource instanceof AbstractModuleSource);

  const toStringTag = Object.getOwnPropertyDescriptor(AbstractModuleSource.prototype, Symbol.toStringTag).get;
  assert_equals(toStringTag.call(exportedNamesSource), "WebAssembly.Module");
}

assert_array_equals(WebAssembly.Module.exports(exportedNamesSource).sort(),
                    ["func", "glob", "mem", "tab"]);

const wasmImportFromWasmSource = await import.source("./resources/wasm-import-from-wasm.wasm");

assert_true(wasmImportFromWasmSource instanceof WebAssembly.Module);

let logged = false;
const instance = await WebAssembly.instantiate(wasmImportFromWasmSource, {
  "./wasm-export-to-wasm.wasm": {
    log () {
      logged = true;
    }
  }
});
instance.exports.logExec();
assert_true(logged);

done();
</script>
