LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/source-phase.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected identifier 'exportedNamesSource'. Expected 'from' before imported module name.

Harness Error (FAIL), message = SyntaxError: Unexpected identifier 'exportedNamesSource'. Expected 'from' before imported module name.
